Extract path parsing
